Ajout de l'algorithme de dessin de triangle, pentagone et hexagone.

+++ Gaudrophone/src/gaudrophone/Domaine/Instrument/Bordure.java
@@ -30,7 +30,7 @@ public class Bordure {
         visible = valeur;
     }
     
-    public float getLargeur()
+    public int getLargeur()
     {
         return largeur;
     }
